% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_included.R
\name{at_least_correct_predictors_included}
\alias{at_least_correct_predictors_included}
\title{Function for displaying the percent of the time the correct model or the
 correct model with additional covariates is selected}
\usage{
at_least_correct_predictors_included(Simulation)
}
\arguments{
\item{Simulation}{Simulation object}
}
\description{
Function for displaying the percent of the time the correct model or the
 correct model with additional covariates is selected
}
