% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_heat_map.R
\name{betas_heat_map}
\alias{betas_heat_map}
\title{Function for plotting fitted betas as a heat map}
\usage{
betas_heat_map(Simulator, hide_intercept = TRUE)
}
\arguments{
\item{Simulator}{Simulator object}

\item{hide_intercept}{Whether to inlcude the intercept in the heat map}
}
\description{
Function for plotting fitted betas as a heat map
}
