% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_heat_map.R
\name{betas_heat_map_helper}
\alias{betas_heat_map_helper}
\title{Helper function for betas_heat_map}
\usage{
betas_heat_map_helper(matrix_slice, algo_name, hide_intercept)
}
\arguments{
\item{matrix_slice}{fitted coefficient values from a single stepwise algorithm}

\item{hide_intercept}{passthrough parameter}
}
\value{
a ggplot of a beta heat map from a single stepwise algorithm
}
\description{
Helper function for betas_heat_map
}
