% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_bias_variance.R
\name{coeff_bias}
\alias{coeff_bias}
\title{Function for displaying fitted coefficient bias in a table}
\usage{
coeff_bias(Simulation, rounding_digits = 3)
}
\arguments{
\item{Simulation}{object}
}
\description{
Function for displaying fitted coefficient bias in a table
}
