% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_bias_variance.R
\name{coeff_variance}
\alias{coeff_variance}
\title{Function for displaying fitted coefficient variance in a table}
\usage{
coeff_variance(Simulation)
}
\arguments{
\item{Simulation}{object}
}
\description{
Function for displaying fitted coefficient variance in a table
}
