% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FromModelResponseCalculator.R
\docType{class}
\name{FromModelResponseCalculator}
\alias{FromModelResponseCalculator}
\title{Class for calculating response using predict of model}
\format{\code{\link{R6Class}} object}
\usage{
FromModelResponseCalculator
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class for calculating response using predict of model
}
\section{Methods}{

\describe{
  \item{\code{new(object_model, irreducible_error_generator, response_is_continuous)}}{Class constructor}
  \item{\code{get_betas}}{Returns the coefficients of the object_model}
  \item{\code{calculate_response(predictors)}}{Calculates the response using prediction on the object}
  \item{\code{response_is_continuous()}}{Returns whether the response is continuous}
  }
}

\keyword{datasets}
