% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inclusion_order.R
\name{inclusion_order}
\alias{inclusion_order}
\title{Function for outputting inclusion order as a table}
\usage{
inclusion_order(Simulation, pdf_head = 10)
}
\arguments{
\item{Simulation}{object}

\item{pdf_head}{if pdf output how many rows to show, at maximum}
}
\description{
Function for outputting inclusion order as a table
}
