% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalPredictorsGenerator.R
\docType{class}
\name{NormalPredictorsGenerator}
\alias{NormalPredictorsGenerator}
\title{Class for generating correlated continuous normal predictors}
\format{\code{\link{R6Class}} object.}
\usage{
NormalPredictorsGenerator
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class for generating correlated continuous normal predictors
}
\section{Fields}{

\describe{
\item{\code{covariance_matrix}}{Private. Covariance matrix used for generating predictors}

\item{\code{num_observations}}{Private. The number of observations}

\item{\code{num_predictors}}{Private. The number of predictors}
}}

\section{Methods}{

\describe{
  \item{\code{new}}{Constructor(num_observations : int,
   num_predictors : int, covariance_matrix : matrix<double>)}
  \item{\code{get_num_observations}}{Returns num_observations}
  \item{\code{get_num_predictors}}{Returns num_predictors}
  \item{\code{simulate_predictors}}{Returns matrix of predictors}
  }
}

\keyword{datasets}
