% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_included.R
\name{only_correct_predictors_included}
\alias{only_correct_predictors_included}
\title{Function for displaying the percent of the time the correct model is selected}
\usage{
only_correct_predictors_included(Simulation)
}
\arguments{
\item{Simulation}{Simulation object}
}
\description{
Function for displaying the percent of the time the correct model is selected
}
