% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictorsGenerator.R
\docType{class}
\name{PredictorsGenerator}
\alias{PredictorsGenerator}
\title{Abstract Class for generating predictors}
\format{\code{\link{R6Class}} object}
\usage{
PredictorsGenerator
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Abstract Class for generating predictors
}
\section{Fields}{

\describe{
\item{\code{cov_mat}}{Private. Covariance matrix used for generating predictors}

\item{\code{num_observations}}{Private. The number of observations}

\item{\code{num_predictors}}{Private. The number of predictors}
}}

\section{Methods}{

\describe{
  \item{\code{new}}{Abstract class}
  \item{\code{get_num_observations}}{Returns num_observations}
  \item{\code{get_num_predictors}}{Returns num_predictors}
  \item{\code{simulate_predictors}}{Abstract function}
  }
}

\keyword{datasets}
