% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportion_included.R
\name{proportion_included}
\alias{proportion_included}
\title{Function for displaying the percent of the time each covariate is included}
\usage{
proportion_included(Simulation)
}
\arguments{
\item{Simulation}{Simulation object}
}
\description{
Function for displaying the percent of the time each covariate is included
}
