% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResponseCalculator.R
\docType{class}
\name{ResponseCalculator}
\alias{ResponseCalculator}
\title{Abstract Class for calculating response}
\format{\code{\link{R6Class}} object}
\usage{
ResponseCalculator
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Abstract Class for calculating response
}
\section{Methods}{

\describe{
  \item{\code{new}}{Abstract class}
  \item{\code{get_betas}}{Abstract function}
  \item{\code{calculate_response}}{Abstract function}
  \item{\code{response_is_continuous}}{Abstract function}
  }
}

\keyword{datasets}
