% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_cor_matrix.R
\name{sample_cor_matrix}
\alias{sample_cor_matrix}
\title{Function for displaying generated sample correlation matrix.}
\usage{
sample_cor_matrix(Simulator)
}
\arguments{
\item{Simulator}{Simulator object}
}
\description{
Function for displaying generated sample correlation matrix.
}
