% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Simulation.R
\docType{class}
\name{Simulation}
\alias{Simulation}
\title{Class for Simulating stepwise}
\format{\code{\link{R6Class}} object}
\usage{
Simulation
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class for Simulating stepwise
}
\section{Fields}{

\describe{
\item{\code{SimulationDataGenerator}}{R6 object for simulating data}

\item{\code{ResponseCalculator}}{object for computing response}
}}

\section{Methods}{

\describe{
  \item{\code{new}}{Constructor}
  \item{\code{simulate}}{Runs stepwise algorithms}
  \item{\code{get_inclusion_orders}}{Returns the order of inclusions}
  \item{\code{get_fitted_coefficients}}{Returns the fitted coefficients}
  \item{\code{get_test_sse}}{Returns the sse on a new test set}
  \item{\code{get_training_sse}}{Returns the sse on the training set}
  \item{\code{get_test_classification_rate}}{Returns the classification rate on a new test set}
  \item{\code{get_training_classification_rate}}{Returns the classification rate on the training set}
  }
}

\keyword{datasets}
