% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step2.R
\docType{class}
\name{Step2}
\alias{Step2}
\title{Class/algorithm for Stepwise regression that considers negative correlation structure (no recursion)}
\format{\code{\link{R6Class}} object}
\usage{
Step2
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Class/algorithm for Stepwise regression that considers negative correlation structure (no recursion)
}
\section{Fields}{

\describe{
\item{\code{PredictorsGenerator}}{object for generating predictors}

\item{\code{ResponseCalculator}}{object for computing response}
}}

\section{Methods}{

\describe{
  \item{\code{new}}{Constructor}
  \item{\code{run}}{Runs the stepwise regression and saves the fitted model and the order the covariates were included}
  \item{\code{get_fitted_model}}{Returns the glm or lm object (must call "run()" first)}
  \item{\code{get_inclusion_order}}{Returns the order covariates were included in (list<vector<string>>) (must call "run()" first)}
  }
}

\keyword{datasets}
