% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StepLmWrapper.R
\docType{class}
\name{StepLmWrapper}
\alias{StepLmWrapper}
\title{Wrapper class for basic Stepwise regression}
\format{\code{\link{R6Class}} object}
\usage{
StepLmWrapper
}
\value{
Object of \code{\link{R6Class}}
}
\description{
Wrapper class for basic Stepwise regression
}
\section{Methods}{

\describe{
  \item{\code{new}}{Constructor}
  \item{\code{run}}{Runs the stepwise regression and saves the fitted model and the order the covariates were included}
  \item{\code{get_fitted_model}}{Returns the glm or lm object (must call "run()" first)}
  \item{\code{get_inclusion_order}}{Returns the order covariates were included in (list<vector<string>>) (must call "run()" first)}
  }
}

\keyword{datasets}
