% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{test_mse2dataframe}
\alias{test_mse2dataframe}
\title{Function for outputting test mse in a dataframe}
\usage{
test_mse2dataframe(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting test mse in a dataframe
}
