% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{test_mse2kable}
\alias{test_mse2kable}
\title{Function for outputting test mse in a knitr kable}
\usage{
test_mse2kable(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting test mse in a knitr kable
}
