% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{test_mse_plotter}
\alias{test_mse_plotter}
\title{Function for plotting test mse}
\usage{
test_mse_plotter(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for plotting test mse
}
