% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{test_mse_tables}
\alias{test_mse_tables}
\title{Function for outputting test mse in a table}
\usage{
test_mse_tables(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting test mse in a table
}
