% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R, R/test_sse_plotter.R
\name{test_sse2dataframe}
\alias{test_sse2dataframe}
\alias{test_sse2dataframe}
\title{Function for outputting test sse in a dataframe}
\usage{
test_sse2dataframe(Simulation)

test_sse2dataframe(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}

\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting test sse in a dataframe

Function for outputting test sse in a dataframe
}
