% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R, R/test_sse_plotter.R
\name{test_sse2kable}
\alias{test_sse2kable}
\alias{test_sse2kable}
\title{Function for outputting test sse in a knitr kable}
\usage{
test_sse2kable(Simulation)

test_sse2kable(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}

\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting test sse in a knitr kable

Function for outputting test sse in a knitr kable
}
