% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R, R/test_sse_plotter.R
\name{test_sse_boxplot}
\alias{test_sse_boxplot}
\alias{test_sse_boxplot}
\title{Function for plotting test sse}
\usage{
test_sse_boxplot(Simulation)

test_sse_boxplot(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}

\item{Simulation}{Simulation object to plot}
}
\description{
Function for plotting test sse

Function for plotting test sse
}
