% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{training_mse2dataframe}
\alias{training_mse2dataframe}
\title{Function for outputting training mse in a dataframe}
\usage{
training_mse2dataframe(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting training mse in a dataframe
}
