% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{training_mse2kable}
\alias{training_mse2kable}
\title{Function for outputting training mse in a knitr kable}
\usage{
training_mse2kable(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting training mse in a knitr kable
}
