% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{training_mse_plotter}
\alias{training_mse_plotter}
\title{Function for plotting training mse}
\usage{
training_mse_plotter(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for plotting training mse
}
