% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_mse_plotter.R
\name{training_mse_tables}
\alias{training_mse_tables}
\title{Function for outputting training mse in a table}
\usage{
training_mse_tables(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting training mse in a table
}
