% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R, R/test_sse_plotter.R
\name{training_sse2dataframe}
\alias{training_sse2dataframe}
\alias{training_sse2dataframe}
\title{Function for outputting training sse in a dataframe}
\usage{
training_sse2dataframe(Simulation)

training_sse2dataframe(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}

\item{Simulation}{Simulation object to plot}
}
\description{
Function for outputting training sse in a dataframe

Function for outputting training sse in a dataframe
}
