% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fit.R, R/test_sse_plotter.R
\name{training_sse_boxplot}
\alias{training_sse_boxplot}
\alias{training_sse_boxplot}
\title{Function for plotting training sse}
\usage{
training_sse_boxplot(Simulation)

training_sse_boxplot(Simulation)
}
\arguments{
\item{Simulation}{Simulation object to plot}

\item{Simulation}{Simulation object to plot}
}
\description{
Function for plotting training sse

Function for plotting training sse
}
